Rename `hooks` back to `useWebSocket`
